[DEV] Add route for Home page
Create a Home component to contain the contain of the home page.
Add a route to app-routing.module.ts so that when the url path after the domain name is null, it will call the HomeComponent.
Add a place holder <router-outlet> in the app.component.html template so that it will load the component corresponding to the path given