[DEV] Clips sharing page
Create a component for the clip sharing page.
Add dynamic routing using the clip id to app-routing module.
Subscribe to params.id in activated route to get the latest id, so that when user click a particular clip, it will be redirected correctly. 

When a user is navigating to a new page, the previous component will be destroyed. But if Angular detected that the user is navigating to the same component (in this case, clips) it will not destroy and reload it to save resources.