[DEV] Add video sorting function
Add sort function to manage which extract the selected element value as 'sort' and place it in the url route as a query parameter.
Subscribe to the route and get the 'sort' parameter continuously from init.
Call the sort function on change event and preserve the selected option by using the [selected] property binding.
Initialize the manage page with the sort option = 1 in the nav bar.