[DEV] Block default events when uploading videos
Create a directive that helps to block default behaviors of the host element when the user dragover and drop contents to the Upload Video area. By default, the host will create a new tap and show the content.

Export the event block directive in the shared module.

import the shared module in the video module.

Add the event blocker directive to the upload template's upload dropbox to apply the block to the division.