[DEV] Reactive form for uploading videos
Import ReactiveFormsModule in video module where upload module is declared.
Create a form group and add the title with validators to it.
Set the title to the video title by default when an mp4 file is dropped.
Leverage app-input component where we can check the validity and feedback a response to the user.
Change the button's visibility based on the validity of the form.
Create a function to upload the file (with the form)