[DEV] Store and add file data to database
Create a model IClip that describe a data describe a clip.
Create a service that helps add the Clips data to the database.
Create a reference for the clip and get the download url by using AngularFireStorage.
Add auth service to the upload component so that we can get the uid and display name of the user.
Add the clip service to the upload component so that we can add data to the database.