[DEV] Update clips
Modify the edit form as a reactive form with clipID and title.
Add OnChanges implementation for edit component so that we can update the clipID and title on changes.
Add submit() to update changes on Firebase.
Add alert component to notify user about the update status.
Output (emit) an event to manage component for updating the list of clips shown to user.