[DEV] Infinite scrolling clips list
Create a clipsList component and move the Clips List template from Home page to it because we are also using it in the Manage page.
In the clipsList component, we listen to the scroll event on init and remove the event listener on destory.
Created a handleScroll function to check the current scroll position.