[DEV] Video player play video
Get data from firebase using Resolve from Router.
Add the resolver to routing so that we can get the data from Firebase before load the component.
In clip component class, retrieve the clip we got when routing as clip and pass it to the video player in ngOnInit().
Add DataPipe as provider so that we can inject it in our fbTimestamp pipe.
Replace the static template with the information in the clip.
Guard fbTimestamp pipe since we can receive a null clip and therefore there can be no timestamp.